rename depth_image_utils and image_utils packages
