+++ depth_image_utils_drs/package.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0"?>
 <package>
-  <name>depth_image_utils</name>
+  <name>depth_image_utils_drs</name>
   <version>0.0.0</version>
 
   <description>Description</description>
@@ -12,10 +12,10 @@
 
   <buildtool_depend>catkin</buildtool_depend>
   <build_depend>cv_bridge</build_depend>
-  <build_depend>image_utils</build_depend>
+  <build_depend>image_utils_drs</build_depend>
   <build_depend>image_transport</build_depend>
   <run_depend>cv_bridge</run_depend>
-  <run_depend>image_utils</run_depend>
+  <run_depend>image_utils_drs</run_depend>
   <run_depend>image_transport</run_depend>
