+++ depth_image_utils_drs/src/depth_image_utils.cpp
@@ -1,4 +1,4 @@
-#include "depth_image_utils/depth_image_utils.hpp"
+#include "depth_image_utils_drs/depth_image_utils.hpp"
 
 #include <sensor_msgs/image_encodings.h>
