+++ depth_image_utils_drs/src/depth_to_pointcloud_node.cpp
@@ -15,7 +15,7 @@
 
 #include <sensor_msgs/PointCloud2.h>
 
-#include <depth_image_utils/depth_image_utils.hpp>
+#include <depth_image_utils_drs/depth_image_utils.hpp>
 
 using namespace std;
