+++ image_utils_drs/package.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0"?>
 <package>
-  <name>image_utils</name>
+  <name>image_utils_drs</name>
   <version>0.0.0</version>
 
   <description>Description</description>
